+++ topothehourbot/utilities/channels/ircv3/__init__.py
@@ -1,1 +1,2 @@
+from .channel import *
 from .command import *
