--- topothehourbot/websocket/channels/__init__.py
@@ -1,1 +1,0 @@
-from .channels import *
